move part of xpu_op_list.h from fluid to phi
