+++ paddle/phi/backends/xpu/CMakeLists.txt
@@ -1,0 +1,1 @@
+cc_library(phi_xpu_op_list SRCS xpu_op_list.cc)
