+++ paddle/phi/backends/xpu/xpu_op_list.h
@@ -1,0 +1,25 @@
+/* Copyright (c) 2022 PaddlePaddle Authors. All Rights Reserved.
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+    http://www.apache.org/licenses/LICENSE-2.0
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License. */
+#pragma once
+
+#ifdef PADDLE_WITH_XPU
+#include <string>
+
+namespace phi {
+namespace backends {
+namespace xpu {
+
+bool is_in_xpu_black_list(const std::string& op_name);
+
+}  // namespace xpu
+}  // namespace backends
+}  // namespace phi
+#endif
