+++ paddle/phi/core/CMakeLists.txt
@@ -19,7 +19,7 @@ if(WITH_XPU)
   cc_library(
     kernel_factory
     SRCS kernel_factory.cc
-    DEPS phi_enforce convert_utils xpu_op_list)
+    DEPS phi_enforce convert_utils phi_xpu_op_list)
 else()
   cc_library(
     kernel_factory
